+++ tests/test_config.rs
@@ -1,5 +1,6 @@
 mod utils;
 use path_server::config;
+use tower_lsp::lsp_types;
 use utils::*;
 
 /// test config.completion.maxResults
@@ -18,6 +19,7 @@ async fn test_config_max_results_limits_items() {
         show_hidden_files: true,
         exclude: vec![],
         base_path: vec!["${workspaceFolder}".into()],
+        trigger_next_completion: true,
     };
     let cfg = config::Config {
         completion: completion_cfg,
@@ -47,6 +49,7 @@ async fn test_config_show_hidden_files_false() {
         show_hidden_files: false,
         exclude: vec![],
         base_path: vec!["${workspaceFolder}".into()],
+        trigger_next_completion: true,
     };
     let cfg = config::Config {
         completion: completion_cfg,
@@ -78,6 +81,7 @@ async fn test_config_exclude() {
         show_hidden_files: true,
         exclude: vec!["*.log".into()],
         base_path: vec!["${workspaceFolder}".into()],
+        trigger_next_completion: true,
     };
     let cfg = config::Config {
         completion: completion_cfg,
@@ -109,6 +113,7 @@ async fn test_config_base_path() {
         show_hidden_files: true,
         exclude: vec![],
         base_path: vec!["${workspaceFolder}/alt".into()],
+        trigger_next_completion: true,
     };
     let cfg = config::Config {
         completion: completion_cfg,
@@ -124,4 +129,68 @@ async fn test_config_base_path() {
     let labels: Vec<String> = items.into_iter().map(|i| i.label).collect();
 
     assert!(labels.contains(&"config.json".to_string()));
+}
+
+/// test config.completion.triggerNextCompletion
+#[tokio::test]
+async fn test_config_trigger_next_completion() {
+    let harness = TestHarness::new().await;
+
+    // test trigger_next_completion == true
+    harness.create_file("dir/file1");
+    harness.create_file("dir/file2");
+
+    let cfg_with = config::Completion {
+        max_results: 0,
+        show_hidden_files: true,
+        exclude: vec![],
+        base_path: vec!["${workspaceFolder}".into()],
+        trigger_next_completion: true,
+    };
+    harness
+        .set_config(config::Config {
+            completion: cfg_with,
+        })
+        .await;
+
+    let content = "let f = ./d";
+    let uri = harness.open_doc("src/main.rs", content).await;
+
+    let items = harness
+        .completion_items(&uri, 0, content.len() as u32)
+        .await;
+    assert_eq!(items.len(), 1);
+    assert_eq!(items[0].insert_text.clone().unwrap(), "dir/");
+    assert_eq!(
+        items[0].command,
+        Some(lsp_types::Command {
+            title: "triggerSuggest".to_string(),
+            command: "editor.action.triggerSuggest".to_string(),
+            arguments: None,
+        })
+    );
+
+    // test trigger_next_completion == false
+    let cfg_without = config::Completion {
+        max_results: 0,
+        show_hidden_files: true,
+        exclude: vec![],
+        base_path: vec!["${workspaceFolder}".into()],
+        trigger_next_completion: false,
+    };
+    harness
+        .set_config(config::Config {
+            completion: cfg_without,
+        })
+        .await;
+
+    let content = "let f = ./d";
+    let uri = harness.open_doc("src/main.rs", content).await;
+
+    let items = harness
+        .completion_items(&uri, 0, content.len() as u32)
+        .await;
+    assert_eq!(items.len(), 1);
+    assert_eq!(items[0].insert_text.clone().unwrap(), "dir");
+    assert_eq!(items[0].command, None);
 }