+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to the Path Server will be documented in this file.
 ## [Unreleased]
 ### Added
 - **Core**: Support automatically trigger next completion after selecting a completion.
+- **Core**: Add config entry: `path-server.completion.triggerNextCompletion`.
 
 ## [0.3.0] 2026-03-06
 ### Fixed